+++ src/main/kotlin/com/epam/brn/upload/csv/parser/iterator/impl/GroupMappingIteratorProvider.kt
@@ -1,6 +1,6 @@
-package com.epam.brn.upload.csv.iterator.impl
+package com.epam.brn.upload.csv.parser.iterator.impl
 
-import com.epam.brn.upload.csv.iterator.MappingIteratorProvider
+import com.epam.brn.upload.csv.parser.iterator.MappingIteratorProvider
 import com.epam.brn.upload.csv.record.GroupRecord
 import com.fasterxml.jackson.databind.MappingIterator
 import com.fasterxml.jackson.dataformat.csv.CsvMapper
