+++ src/test/kotlin/com/epam/brn/upload/CsvUploadServiceTest.kt
@@ -4,7 +4,7 @@ import com.epam.brn.repo.ExerciseGroupRepository
 import com.epam.brn.repo.ExerciseRepository
 import com.epam.brn.repo.SeriesRepository
 import com.epam.brn.repo.TaskRepository
-import com.epam.brn.upload.csv.CsvParser
+import com.epam.brn.upload.csv.parser.CsvParser
 import java.io.IOException
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Assertions.assertThrows
